added typings to package.json
