+++ src/responseCollection.ts
@@ -1,10 +1,5 @@
-import getResponseObj, { ResponseObject } from './util'
-
-export interface ResponseCollection
-{
-  local : ResponseObject
-  session : ResponseObject
-}
+import getResponseObj from './util'
+import { ResponseCollection } from './index'
 
 export default function(request$, runStreamAdapter) : ResponseCollection {
   return {
